--- action.go
@@ -1,49 +1,0 @@
-package gb
-
-// Actions and Tasks.
-//
-// Actions and Tasks allow gb to separate the role of describing the
-// order in which work will be done, from describing that work itself.
-// Actions are the former, they describe the graph of dependencies
-// between actions, and thus the work to be done. By traversing the action
-// graph, we can do the work, execute the Tasks in a sane order.
-//
-// Tasks describe the work to be done, without being concerned with
-// the order in which the work is done -- that is up to the code that
-// places Tasks into actions. Tasks also know more intimate details about
-// filesystems, processes, file lists, etc that Actions do not.
-//
-// Action graphs (they are not strictly trees as branchs converge on base actions)
-// contain only work to be performed, there are no Actions with empty Tasks
-// or Tasks which do no work.
-//
-// Actions are executed by Executors, but can also be transformed, mutated,
-// or even graphed.
-
-// An Action describes a task to be performed and a set
-// of Actions that task depends on.
-type Action struct {
-
-	// Name describes the action.
-	Name string
-
-	// Deps identifies the Actions that this Action depends.
-	Deps []*Action
-
-	// Task identifies the that this action represents.
-	Task
-}
-
-// Task represents some work to be performed. It contains a single method
-// Run, which is expected to be executed at most once.
-type Task interface {
-
-	// Run will initiate the work that this task represents and
-	// block until the work is complete.
-	Run() error
-}
-
-// TaskFn is a Task that can execute itself.
-type TaskFn func() error
-
-func (fn TaskFn) Run() error { return fn() }
